Move XMLConfig to src/ewatercycle/parametersetdb/config.py + move lisflood version image to own module
