+++ tests/models/test_lisflood.py
@@ -12,8 +12,9 @@ from numpy.testing import assert_array_equal
 
 from ewatercycle import CFG
 from ewatercycle.forcing import load_foreign
-from ewatercycle.models.lisflood import Lisflood, XmlConfig
+from ewatercycle.models.lisflood import Lisflood
 from ewatercycle.parameter_sets import ParameterSet, example_parameter_sets
+from ewatercycle.parametersetdb.config import XmlConfig
 
 
 @pytest.fixture
